[MCOBERTURA-75] implement skip handling
* add a property to skip the mojo executions
* exclude parent-pom projects from mojo execution by default
* add a forceMojoExecution in case it's still needed 


git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/cobertura-maven-plugin@12173 52ab4f32-60fc-0310-b215-8acea882cd1b

+++ src/main/java/org/codehaus/mojo/cobertura/CoberturaCleanMojo.java
@@ -35,6 +35,11 @@ public class CoberturaCleanMojo
     public void execute()
         throws MojoExecutionException, MojoFailureException
     {
+        if ( skipMojo() )
+        {
+            return;
+        }
+        
         if ( dataFile.exists() )
         {
             dataFile.delete();
